Remove anchor attribute from ValidationNotification
The ValidationNotification component is rendered as a p element
instead of Material UI’s Popover component, so anchor is no longer
a valid attribute. Do not set the anchor attribute when rendering
the component, as it seems to be redundant anyway. The errors are
rendered correctly and the linking from the validation error message
to the errors seems to be working too.